+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Bug fix: https://github.com/p4535992/foundryvtt-pin-cushion/issues/72
 - Bug fix: https://github.com/p4535992/foundryvtt-pin-cushion/issues/67
 - Add feature suggested from https://github.com/foundryvtt/foundryvtt/issues/8770
+- Check for: https://github.com/p4535992/foundryvtt-pin-cushion/issues/66
+- Add feature: https://github.com/p4535992/foundryvtt-pin-cushion/issues/60
 
 ### 1.8.32