+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Check for: https://github.com/p4535992/foundryvtt-pin-cushion/issues/66
 - Add feature: https://github.com/p4535992/foundryvtt-pin-cushion/issues/60
 - Add embedded integration from https://github.com/aMediocreDad/jal for try to solve issue https://github.com/p4535992/foundryvtt-pin-cushion/issues/64
+- Bug fix: https://github.com/p4535992/foundryvtt-pin-cushion/issues/69 added "Page Section" field
 
 ### 1.8.32